requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
google-auth==2.20.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gspread==5.9.0
idna==3.4
IMAPClient==2.3.1
magic-filter==1.0.9
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==1.26.16
xlrd==1.2.0
yarl==1.9.2
